chore(frontend-entries): clean-up and bringing components out to different files
