+++ frontend/evoplanner_frontend/components/ScheduleColumn.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ScheduleSlot from "./ScheduleSlot";
 import DragPreview from "./DragPreview";
-import { calculateSlotPosition } from "@/utils/schedule";
+import { calculateSlotPosition } from "@/utils/scheduleDisplay";
 
 export default function ScheduleColumn({
   day,
